# server/src/openapi.yaml
openapi: 3.0.3
info:
  title: Book Review App API
  description: |
    書籍レビューアプリケーションのREST API仕様

    🔐 認証について：
    - 保護されたエンドポイント（🔒マーク）には Authorization ヘッダーが必要です
    - Format: `Authorization: Bearer <JWT_TOKEN>`
  version: 1.0.0
  contact:
    name: Book Review App
  license:
    name: ISC
servers:
  - url: http://localhost:3001/api
    description: ローカル開発環境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 吾輩は猫である
        author:
          type: string
          example: 夏目漱石
        publicationYear:
          type: integer
          example: 1905
        ISBN:
          type: string
          example: 9784101010014
        summary:
          type: string
        averageRating:
          type: number
          format: float
          example: 4.2
        reviewCount:
          type: integer
          example: 15
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
...
paths:
...
  /books:
    get:
      summary: 書籍一覧取得
      tags: [Books]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
          description: タイトルまたは著者で検索
        - name: minRating
          in: query
          schema:
            type: number
        - name: maxRating
          in: query
          schema:
            type: number
        - name: yearFrom
          in: query
          schema:
            type: integer
        - name: yearTo
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, author, publicationYear, rating, createdAt]
        - name: order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        '200':
          description: 書籍一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      books:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      summary: 書籍追加
      tags: [Books]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
              properties:
                title:
                  type: string
                author:
                  type: string
                publicationYear:
                  type: integer
                ISBN:
                  type: string
                summary:
                  type: string
      responses:
        '201':
          description: 書籍追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Book'
        '400':
          description: バリデーションエラー
        '401':
          description: 未認証
        '409':
          description: ISBN既存
  /books/{id}:
    get:
      summary: 書籍詳細取得
      tags: [Books]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 書籍詳細取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Book'
        '404':
          description: 書籍が見つからない

    put:
      summary: 書籍更新
      tags: [Books]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                publicationYear:
                  type: integer
                ISBN:
                  type: string
                summary:
                  type: string
      responses:
        '200':
          description: 書籍更新成功
        '403':
          description: 権限不足
        '404':
          description: 書籍が見つからない

    delete:
      summary: 書籍削除
      tags: [Books]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 書籍削除成功
        '403':
          description: 権限不足
        '404':
          description: 書籍が見つからない
...